refactor: divide function tests into new file
